Add functionality using a closure
Even though the previous version was clearly superior, I have refactored
the code to use a closure because I can. I have no idea if I can really
utilize a factory function for this or if it even makes sense but good
lord I am going to try my best.